requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
idna==2.6
numpy==1.14.0
Pillow==5.0.0
pytz==2017.3
requests==2.18.4
scipy==1.0.0
urllib3==1.22
